AWS-EC2 BugFix Missing publicIp in Terraform Output
When restarting an EC2 instance, the public IP can change by default. To prevent Terraform from recreating the instance due to a public IP change, we use ignore_changes for the publicIp attribute in the resource configuration.

However, after applying this workaround, the publicIp is sometimes not displayed in the Terraform output. This happens because Terraform no longer tracks changes to the publicIp attribute, and the value may be absent or outdated in the state file

Previously, we relied on Terraform outputs to retrieve instance and volume information. To overcome this issue and always fetch up-to-date details (especially the current public IP), we are now switching to use the AWS SDK for JavaScript (@aws-sdk/client-ec2) to fetch EC2 instance information directly through API calls, rather than relying on Terraform state or outputs.